+++ app/src/main/java/com/example/accountbook/repository/CategoryRepository.kt
@@ -10,8 +10,8 @@ interface CategoryRepository {
     suspend fun update(category: Category)
     suspend fun delete(category: Category)
     suspend fun delete(array: Array<Category>)
-    suspend fun getCategory(id: Int): Category
-    suspend fun getChildrenCategories(pid: Int): Flow<List<Category>>
+    suspend fun getCategory(id: Long): Category
+    suspend fun getChildrenCategories(pid: Long): Flow<List<Category>>
 
     suspend fun getCategories(depth: Int): Flow<List<Category>>
 }
@@ -21,7 +21,7 @@ class CategoryRepositoryImpl(private val db : AppRoomDatabase): CategoryReposito
     override suspend fun update(category: Category) = db.categoryDao().update(category)
     override suspend fun delete(category: Category) = db.categoryDao().delete(category)
     override suspend fun delete(array: Array<Category>) = db.categoryDao().delete(*array)
-    override suspend fun getCategory(id: Int) = db.categoryDao().get(id)
-    override suspend fun getChildrenCategories(pid: Int): Flow<List<Category>> = db.categoryDao().getChildrenCategory(pid)
+    override suspend fun getCategory(id: Long) = db.categoryDao().get(id)
+    override suspend fun getChildrenCategories(pid: Long): Flow<List<Category>> = db.categoryDao().getChildrenCategory(pid)
     override suspend fun getCategories(depth: Int): Flow<List<Category>> = db.categoryDao().getCategories(depth)
 }
